pyproject.toml
[project]
name = "prometeo-mcp"
version = "0.1.0"
description = "Prometeo MCP server interface for testing and integration with LLMs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Prometeo API", email = "dev@prometeoapi.com" }]
dependencies = [
    "pydantic<=2.11",
    "mcp[cli]<=1.8.1",
    "prometeo==2.0.5",
    "python-dotenv<=1.1.0",
    "pyyaml<=6.0.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.23", "anyio>=4.0"]

[build-system]
requires = ["uv", "setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["prometeo_mcp"]